requirements.txt
absl-py==2.3.1
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dlib
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.5.2
keras==3.9.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.1.0
narwhals==2.12.0
networkx==3.5
numpy
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
pluggy==1.6.0
protobuf==4.25.8
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
sympy==1.14.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
